Map all yang ints to int, add unions & identityref
Unless we actively care about size of ints (not super important at this
stage) we might as well use the single int type. We can track allowed
size as range instead (later...)

Decided to use val for unions. It should obviously be mapped to Acton
unions in the end, but we don't have those yet, so we'll map to 'val'
for now. 'val' is like a builtin union of everything, so it works but
won't be elegant nor easy to use.. YANG unions can be a union of any
other YANG type, so we need to allow any of the equivalent Acton types.
Right now we only use builtin Acton types, so we could possibly narrow
down from val to atom, but given that the bits type and maybe
enumeration are likely to need special classes, we likely have to stick
to using 'val'.

Think we can use str for identityref.