Map same resolved union base types to single type
When all base types of a union type are the same type, we use that
type. For example, the ietf-inet-types has the following typedefs

  typedef ipv4-address {
    type string;
  }
  typedef ipv6-address {
    type string;
  }
  typedef domain-name {
    type string;
  }
  typedef ip-address {
    type union {
      type ipv4-address;
      type ipv6-address;
    }
  }
  typedef host {
    type union {
      type ip-address;
      type domain-name;
    }
  }

ip-address is a union of ipv4-address and ipv6-address, which are both
strings, so we can map it to the Acton type string and just have to keep
track of the XOR patterns for ipv4-address and ipv6-address
respectively. The host type is the same but one level deeper.