Add yang.compile()
First, we now have the yang module and not just submodules like
yang.schema and yang.parser. Heads-up: This requires a rather fresh
version of Acton.

`yang.compile(list[str])` takes a list of yang modules sources, i.e.
just read in the source of a bunch of .yang modules feed that in. It
will then compile this as a bundle of YANG module and return a DRoot,
i.e. a YANG tree built from such a collection of modules. Nice and
convenient!