Resolve union of ints as ints
I have prepared some code to do rather nice mapping, so if we have a
group of int sizes (why anyone would ever create a union of that, I do
not know...) then we return the largest int size, like i32 or i64. If we
on the other hand have uints, you get the largest uint. With a mix of
int and uints, we need a signed integer, so we figure out the smallest
one able to hold the mix of values or potentially revert to "int" if
there is a uint64 involved.

However, since we don't have integer subtyping, we just shortcut to
"int" for now in all these cases, but we're ready to change this a
little later...