Add basic submodule support
Just move over the submodule content over into the main module.

I think there are more complex scenarios where we need to rewrite the
content of the submodule to account for a local prefix used that is
different from the main module. Similarly, a submodule can include other
submodules, but in YANG1.1 I think that's equivalent as all submodules
already should see the content of all other submodules under the same
main module, or?