Move gen to separate project dir
We generate the lion's share of schema.act from the YANG RFC. I am now
moving this to a separate directory to make it more clear that it's not
part of the "normal" project. We keep the generated code checked in, so
this doesn't affect how this project is built as a dependency by other
projects. If anything, this is cleaner since we remove files that really
aren't part of the yang module, like rfcgen itself.