Add test for data classes
The generated data classes are particularly hard to test since they are
generated code and so we need multiple passes. To achieve that, we
create separate repos so that we can cobble things together with a
Makefile or similar (until Acton build is customizable!), but at least
we can get some test coverage of our generate from_xml etc functions.