Fix XML parsing of optional nodes
The generated from_xml() methods were a little too eager and required
all nodes to be present, even non-presence containers without any
mandatory children. This was simple the wrong behavior.

We now allow this by inspecting whether the subtree is optional to take
an optional or non-optional node from the XML input. Further, depending
on if a class has required args or not, the from_xml function can return
a default object or raises an exception if it requires input arguments.