Add test case for mixed req / opt args
In the generated class for the list we need to order arguments
correctly. Acton requires non-optional arguments to come before optional
arguments. This is a slightly trickier situation than normally, first we
have the list key which is obviously required, then comes the foo leaf,
which is mapped to a non-optional str since it has a default value,
however, the input argument to the object constructor is an *optional*
str since we will apply the default value if one is not supplied. Thus
we must differentiate between the type in the class vs the constructur
argument type.

Last comes a container bar which is non-optional since it has a
mandatory arg. The generated code is currently wrong since the generated
argument order has the optional foo before the required bar.