Separate class attr type from constructor arg type
We need to compute the type for class attribute separately from the arg
type used in the constructor. For YANG leaves with a default value, we
map it to a non-optional type whereas the constructor argument type
should be optional. Since Acton requires mandatory arguments to come
before optional arguments, we have to keep these apart so we can order
them correctly.

Fixes the mixed_req_args test case.