Add SchemaNode.get_attrs() to expose attrs
Exposing a list of SchemaNode attributes makes it possible to iterate
over its attributes at runtime. This is for example useful when we need
to reconstruct Statement in SchemaNode.to_statement().

But, there is a catch! The list of attributes must be generated on
demand, not in SchemaNode.__init__(), so that we avoid it containing
stale attribute values if the SchemaNode is later transformed.