Direct descendants of choice are optional
The data nodes (DLeaf, DAnydata, DAnyxml) that are direct descendants of
choice/case and are marked as mandatory in the schema are made optional
as DNode's. We don't currently have a way of defining (mutually
exclusive) groups of constraints on DNode's so we just disable these for
the time being.