Add prefix to DNode, prepend to conflicting children
If we detect children with conflicting names in DNode.prdaclass() then
we prefix both attributes, names (in adata, gdata) with the module
prefix. In doing so each node is addressable with an unique name. For
example, a node named "c1" defined in modules "foo" and "bar" at the
same level (data node or augment) becomes:
- "foo:c1" and "bar:c1" in gdata
- "foo_c1" and "bar_c1" in adata

We strip the prefix and separator when printing gdata.
We use a single underscore as the separator in adata to avoid (Acton
bug?) class attribute and class name collision?!