Expose intermediate schema compilation results
The new
- yang.compile_modules(yang_sources: list[str]) -> list[yang.schema.Module]
- yang.compile_to_dmodule(modules: list[yang.schema.Module) -> list[yang.schema.DModule]

expose intermediate compilation result, before combining all modules into a
single root. This is better suited for printing the tree representation of a
module because we get access to SchemaNode (Module) after children expansion.
The module information is discarded in the list[DModule] -> DRoot conversion so
the final DRoot is not the best candidate for printing the schema.