Include namespaces in generated Acton modules
The YANG namespaces of the modules we compiled from, or rather, the
namespaces that made it into nodes in the final tree are now included as
schema_namespaces constant in the generated Acton module. This can be
useful to latest inspect what namespaces we can expect from a particular
schema.