Replace gdata.Module with Container
It means the pattern that we used to have where we could create a Module
at the top and set its namespace is no longer possible or valid. The top
node should be thought of just as a pseudo-node regardless if it is a
Root (what we used to have for content of multiple modules) or
Module (i.e. a single YANG module). That top root node should never have
a namespace, it is the second level nodes in under the root that should
ALWAYS have a namespace set.