Merge gdata.ListElement with Container
The content of a list element is essentially the same as a Container,
just a bunch of children. The biggest difference is that a ListElement
also has a key. Keeping track of these separately in diff, merge etc is
a lot of work, so we now merge them. Container is the new name that gets
used everywhere and it does support an optional key concept.