Simplify gdata
Remove AbsentListElement and let Absent represent this use case as well.

Given the earlier merge and removal of Root, Module etc, Container is
now the only Inner node, thus we have dropped Inner and replaced it with
Container in all those uses.

The class hierarchy isn't as "clean", so it can now represent data trees
that are fundamentally invalid, but at the same time, this class
hierarchy simplifies so many other things, so we just have to keep track
of and do the right thing in code. gdata is an internal representation,
that, except for a few test cases, should only be constructed by code we
control, in adata etc. For an example of "less clean", since Absent can
stand in for any Node, like a Leaf or Container, it inherits directly
from Node. Since Absent needs to stand in for Container, which can have
keys, it means also Absent needs keys, and so .key is now an attribute
on Node. This is nonsensical for all the other uses of Node or its
descendents like Leaf which never have a key but it does allow us to
have simpler merge / diff / patch etc algorithms.