Read key value directly from key leaf
The Container.key attribute is a string list of keys for the list entry.
The way the string values for each leaf key are constructed is internal
to gdata and is independent of the output output format. We must look up
the actual value object from the child leaf and then apply the correct
formatting for the chosen output.