Container.key_str to read values from key leafs
Create the "key" identifier by reading the key values directly from the
list element Container. Because only the List object knows the names of
the keys, it passes the list of names to the method.

Include key leafs in diff for list element Container nodes. Though they
are by definition equal for a modified list element, they must be
present to identify the list element.

Remove Node.key attribute. As a side effect, the Container nodes are no
longer orderable.