Test gdata source roundtrip
In test_data_classes we now also store the gdata test output as Acton
code in the test_gdata_source_roundtrip project. This makes
test_gdata_source_roundtrip the third project in the chain. We restore
the gdata from code, dump it out to XML again to check whether we get
the same document that we used as input.