Move validation closer to XML, JSON val takers
In this refactoring of the XML, JSON deserialization code we move the
validation logic for special cases closer to the input. The special
cases are empty leaves and 64-bit numbers, where the Acton type is
different as to how the value appears in XML, JSON.

As a result the generated code in adata modules is now simpler, and
could be further refactored to extract common patterns for XML, JSON
deserialization.