Include schema in adata modules
The schema is used by the schema-driven XML and JSON deserializers
Ideally this would be the compiled schema (DNode tree), but that is
currently not possible to serialize cleanly due to the parent attribute
adding a cyclic dependency in the object graph. Until we refactor DNode
to support a more declarative style, we store the input YANG sources
instead.