Require fully qualified Identityref in adata
Refactor identityref handling to require complete Identityref values in
generated data classes (adata). This naturally encourages the users to
use generated module constants - singleton Identityref objects generated
for all defined identities. Referencing one of these constants is as
simple as:

```
c.l_identityref = yang_foo.foo_id
```

Internally the XML and JSON value takers return PartialIdentityref
instances. That is necessary because the namespace qualifiers for an
identity are optional if the identity is local to the module, and
beacuse we only get partial information depending on the source data
format. These partial instances are then transformed to a complete
Identityref as part of conversion to gdata and are validated to conform
to the identityref type derivation rules.

As a result, we now store identityref leaves in gdata with complete
namespace information. Perhaps we can later refactor the gdata -> XML,
JSON serializers to skip the namespace qualifiers where they are
unnecessary.