Support eq on different types
For YANG leaf union types, the concrete values we get might be of
different types, like a string and some integer. It is valid to compare
them, for example for the purpose of computing a diff, and the result is
always False. For diff computation this means that the values are always
seen as different and thus the resulting diff will overwrite it - which
is what we want.

Previously we considered comparing different types a sort of developer
bug that should probably warrant more investigation. We loose that
ability now. Not sure how useful it was / is. An alternative would be to
thread through the supposed type of yang leafs, so we would see "union"
and thus only allow the equality check of different concrete types when
the leaf is a union... but meh, keep it simple for now.