Add DInput, DOutput to DRpc.children and attributes
If the RPC Input or Output in SchemaNode do not include any data node
children, do not include them in DRpc when converting to DInput and
DOutput. This simplifies the adata generator a bit. We still set the
SchemaNode Rpc.input/output attributes in Rpc.compile() implicitly
because it makes easier to support augmentation (lookup of
.../foo-rpc/input always succeeds).

In DRpc we then expose DInput and DOutput via:
- the "children" attribute, because this makes the adata code generator
  simpler - it just iterates over the DRpc.children as it would with any
  other DNodeInner.
- the "input" and "output" optional attributes, because this makes the
  rpc_root wrappers code generator simpler - it has readily available info
  whether the RPC has defined input / output.