Add txid
Transaction ID (txid) is a way to track a particular version or ID of
configuration. Using it, we can compare configuration equality in a very
cheap way, if the txids are the same, the config is the same (or should
be).

https://datatracker.ietf.org/doc/draft-ietf-netconf-transaction-id/
specifies a general mechanism for the use of transaction id. Many
devices today have non-standard support for some form of metadata
information that can be used as txid. For example on JUNOS returns a
"commit-seconds" XML attribute for get-config which can be thought of as
a txid for the whole configuration.

We want to be able to synchronize individual subtrees which is why txid
is placed on Node, so that there can be one txid for the top-node and
further down the tree we have other txids for each subtree.