Place path prefix remapping into separate function
The SchemaNode.update_namespace_qualifiers now only overwrites the node
and its children namespace qualifiers, which we must do when placing
submodule children into the parent module.

The SchemaNode.remap_prefix_references updates the node and its children
references to an imported module's own prefix, not local. This
transformation applies to modules too, if a module imports another
module with a prefix different from the one defined in the imported module.