Use empty string for unnamed top-level node
We use the name argument to pass the current node name - in gdata the
name is defined in the containing parent node. An empty string is a
special case that denotes the top level.

Technically a valid XML document only has a single top-level node. The
Node.to_xmlstr() function violates this as it may produce multiple
top-level nodes.