Add gdata.Node.to_xml() -> list[xml.Node]
We now convert gdata directly to xml.Node. This makes the output data
easier to use in the NETCONF client, which works with xml.Node, not XML
strings. We keep the Node.to_xmlstr() method for logging, but XML
encoding is now offloaded to the xml module, where special characters
are encoded correctly.