Expose only DRoot.prdaclass()
We hide the recursion arguments and accumulators in an internal
_prdaclass_rec() method. Just like the gen3 parsers,
DNode._prdaclass_rec accumulates the list of visited DNodes in spath
argument. This is then used for printing the path in
get_path(list[DNode]) -> str.

Even with .prdaclass() only being available on DRoot, it is still
possible to print adata for a schema subtree, like a filter. With the
root_path: list[str] argument it will first accumulate the schema path
to the target node and start printing classes from there. Because this
schema path is then used to compute class names and such, all the names
in the filtered output are equal to unfiltered.