Use Present type for empty leaf gdata value
This is a step towards removing the gdata.Node.type attribute - we
differentiate between boolean and empty type leaf values. Ideally the
Present instance would not carry additional properties (present
attribute) and we would instead rely on the presence of an instance.
This would make the value attribute type optional (?Present) and there
are other consequences to that, so it will be done at a later time.