Implement DNode.prsrc()
With the DNode.parent leaf gone, we can print the entire DNode tree in a
single expression! The only remaining wart is Type object, still of
SchemaNode base class. Its initializer has the mut effect because it
sets the parent attribute on its child attributes. But we're working on
replacing those with DType very soon. With pure initializers it will be
possible to print the entire DNode tree as a module constant.