Implement YangData for yang.adata.MNode
Conversion from adata objects to gdata is now done with the same gen3
schema-driven parsers as the other formats (XML, JSON). The benefit of
this approach is that we only define the rules for validation (ranges,
other limitations) and correct type conversion (unions) once - in
try_parse_value.

To achieve this, we implement the YangData protocol on the base type
MNode. To get the attribute values at runtime, each adata class has a
_get_attr(str) -> ?value method with a lookup table mapping the unique
safe name of an attribute to its value.

Each adata class still includes a convenience to_gdata() method, but it
is just a wrapper on the gen3.from_data() function that gets the adata
instance and the root path.