Adapt to Acton builtin-i64 branch
All numeric values in gdata are now explict bigints. This is essentially
the same behavior as before, when "int" was "bigint". Now "int" is i64,
so we remain with the same semantics by switching to the type "bigint".

Going a little overboard with conversion of negative i64
-9223372036854775808 but I actually saw a crash from one occurrence. I
want to fix proper parsing before digging deeper into those, but we
really should be using the literal -9223372036854775808 and not the
workaround -9223372036854775807-1..