Thread parent DList where needed
The fact that a leaf is part of a key in a list is important when
deciding whether a leaf is optional or non-optional. This affects the
Acton value type in adata, as well as the parsers. By threading the
parent DList node we ensure this information is available where needed
and remove the reliance on the DNode.parent attribute.