WIP: include DNode tree in adata module
Include the printed DNode tree as code in the generated adata module.
The plan is to define it as a module constant so that it avoids the GC
heap. That is not possible just yet because the Type class initializer
has a mut effect. Once we switch over to DType it should work!

Then we can refactor the gen3 wrappers to use the compiled schema
instead of compiling on every call. And if that works, we can remove the
gen2 parsers altogether.