Refactor get_path_name to use list[DNode] path
The function to compute an adata class name given a path to a DNode now
accepts the entire list of DNode nodes, instead of traversing the nodes
in the direction of the root using the parent attribute. With that, the
DNode.parent attribute can be removed!

A neat side effect of this change is that we now apply the rules for
avoiding naming conflicts consistently for all nodes, including
top-level nodes - the module prefix is always used to disambiguate the
conflicting names.