Include DNode tree in adata module
Print the DNode tree as code in the generated adata module. The plan is
to print it as a module constant, which then avoids the GC heap. But
that is not possible today because the Type.__init__ has a mut effect.
The Type class is still of SchemNode kind. I am resonably certain I know
of where the mut effect is coming from - the initializer sets the parent
attribute on children of Type, like Range, ... Anyway, Johan is
currently working on replacing Type with DType classes in the compiled
schema tree, and those have pure initializers. When that work is done we
really can move the source to a module constant!