YangData abstract namespace qualifier implementation
When we do a data lookup from YangData (XML, JSON), the rules for a
namespace qualified lookup are aligned - the top node must be namespace
qualified, as well as children that belong to a different YANG module.
But the namespace qualifier value is different - for XML it is the YANG
module namespace, for JSON it is the YANG module name. The nq_qualified
boolean is now just a signal that a qualified lookup is required.