Remove namespace qualifiers from DType
After a DType has been resolved an materialized, then it doesn't really
matter which module defined it. The same instance may be shared across
modules as long as the input restrictions are the same, which is what we
hash anyway in ResolvedType. The ResolvedType instances are already
stripped of namespace qualifiers.