Print SRC_DNODE as module constant
With the combined recent changes of removing DNode.parent and replacing
Type(SchemaNode) with DType instance, all DNode initializers now have
the pure effect. This means the entire DRoot tree can be printed as a
module constant! In theory this should reduce pressure on GC because
module constants are allocated and loaded outside of the GC heap.