Add op to YangData.take_leaflist
The way we "merge" the operations for leaf-list is a bit of a hack right
now. We don't have a way of expressing operations on individual
leaf-list elements in gdata, we always replace the entire set of
elements. If the operations are a mix of setting and deleting, collect
the entries that were created and drop the ones that were deleted.
Otherwise we only support a single operation type for all elements.

We plan on refactoring leaf-list handling to be more list-like. That
will enable us to then carry the operations for individual elements
separately in gdata as well.