Add identityref with no possible value
An identityref with a base identity that is not used as a base for any
other identities has no valid value. As per RFC, the identityref value
must be derived from (all) bases; and base is not derived from itself.

We include this in the test_data_classes test to ensure the generated
adata module compiles.