Use qualified name for gdata Node names
gdata node names in the children dict now use qualified names - a
combination of prefix and name so that we may compare gdata trees
produced by different schemas, where one schema is a subset of the
other.

The gdata.Id wrapper class is a workaround for lack of Hashable protocol
implementations on named tuples in Acton. Once the language is improved,
we may replace the class with a named tuple like (q: str, name: str)